const express = require('express');: Imports the Express framework.
const router = express.Router();: Creates a new router object. This is like a mini-Express application that we can use to group our authentication-related routes.

router.post('/register',...): This tells the router to listen for POST requests at the /register path. A POST request is used when a client wants to send data to the server to create something new, like a new user account.

res.send(...): For now, when a request hits this endpoint, we simply send back the text "User registration route".

module.exports = router;: We export the router so that our main server.js file can use it.